Separated out the unit-test for CurrentContent.
Added check on value-resolver to prevent circular-references.
