+++ tests/Our.Umbraco.Ditto.Tests/Our.Umbraco.Ditto.Tests.csproj
@@ -212,11 +212,11 @@
   </Choose>
   <ItemGroup>
     <Compile Include="Attributes\MockValueAttribute.cs" />
+    <Compile Include="CurrentContentTests.cs" />
     <Compile Include="Mocks\PropertyBuilder.cs" />
     <Compile Include="Mocks\ContentBuilder.cs" />
     <Compile Include="Mocks\PublishedContentMock.cs" />
     <Compile Include="Mocks\PublishedContentPropertyMock.cs" />
-    <Compile Include="Models\CurrentContentTestModel.cs" />
     <Compile Include="Models\ComplexModel.cs" />
     <Compile Include="Models\SimpleModel.cs" />
     <Compile Include="PublishedContentTests.cs" />
